+++ samples/13_Calculation.php
@@ -7,7 +7,7 @@ require __DIR__ . '/Header.php';
 // List functions
 $helper->log('List implemented functions');
 $calc = \PhpOffice\PhpSpreadsheet\Calculation::getInstance();
-print_r($calc->listFunctionNames());
+print_r($calc->getImplementedFunctionNames());
 
 // Create new Spreadsheet object
 $helper->log('Create new Spreadsheet object');